ath79: fix TL-WA901ND PHY stuck in reset
Some newer bootloaders for the TP-Link TL-WA901ND put the ethernet PHY
in reset before loading the kernel, thus the LAN interface is not
working in OpenWrt.

The PHY reset pin is controlled via the platforms reset register.
Backport necessary patches and alter the devices DTS to allow the kernel
to bring the PHY out of reset.

This requires quite some patches for kernel 4.14, however all patches
are upstreamed.

Signed-off-by: David Bauer <mail@david-bauer.net>